Create README.md
We believe in democracy, collaboration and innovation and making it easy for people to make trusted decisions in this era of information. The process of exercising our democracy should not therefore be strenuous. The idea came up to me when my friends were having challenges with the applications of a post in the student governing council in Rongo University.